Split ExecuteResult.scala
Aiming at moving the ExecuteResult ADT to the interpreter-api module
